+++ __tests-pacts__/serlo.org/index.ts
@@ -51,5 +51,5 @@ describe('GET /api/uuid/:id', () => {
   require('./uuid')
 })
 describe('GET /api/threads/:id', () => {
-  require('./threads')
+  require('./uuid/thread')
 })
