+++ __tests__/schema/uuid/thread.ts
@@ -23,18 +23,19 @@ import { gql } from 'apollo-server'
 import { rest } from 'msw'
 import * as R from 'ramda'
 
-import { article, user } from '../../__fixtures__/uuid'
-import { comment1, comment2, comment3 } from '../../__fixtures__/uuid/comment'
-import { Service } from '../../src/graphql/schema/types'
-import { UuidPayload } from '../../src/graphql/schema/uuid/abstract-uuid'
-import { CommentPayload } from '../../src/graphql/schema/uuid/thread'
-import { Instance } from '../../src/types'
+import { article, comment2, comment3, user } from '../../../__fixtures__/uuid'
+import { comment1 } from '../../../__fixtures__/uuid/thread'
+import { Service } from '../../../src/graphql/schema/types'
+import { UuidPayload } from '../../../src/graphql/schema/uuid/abstract-uuid'
+import { CommentPayload } from '../../../src/graphql/schema/uuid/thread'
+import { Instance } from '../../../src/types'
 import {
   assertSuccessfulGraphQLQuery,
   Client,
   createJsonHandler,
   createTestClient,
-} from '../__utils__'
+  createUuidHandler,
+} from '../../__utils__'
 
 let client: Client
 
@@ -43,6 +44,50 @@ beforeEach(() => {
     service: Service.SerloCloudflareWorker,
     user: user.id,
   }).client
+})
+
+test('property "createdAt"', async () => {
+  global.server.use(createUuidHandler(comment1))
+  await assertSuccessfulGraphQLQuery({
+    query: gql`
+      query propertyCreatedAt($id: Int!) {
+        uuid(id: $id) {
+          ... on Comment {
+            createdAt
+          }
+        }
+      }
+    `,
+    variables: { id: comment1.id },
+    data: {
+      uuid: { createdAt: comment1.date },
+    },
+    client,
+  })
+})
+
+test('Test property "author"', async () => {
+  global.server.use(createUuidHandler(comment1))
+  global.server.use(createUuidHandler(user))
+
+  await assertSuccessfulGraphQLQuery({
+    query: gql`
+      query propertyCreatedAt($id: Int!) {
+        uuid(id: $id) {
+          ... on Comment {
+            author {
+              username
+            }
+          }
+        }
+      }
+    `,
+    variables: { id: comment1.id },
+    data: {
+      uuid: { author: { username: user.username } },
+    },
+    client,
+  })
 })
 
 test('Threads, 3 Comments', async () => {
