+++ src/graphql/schema/index.ts
@@ -27,7 +27,6 @@ import { jsonSchema } from './json'
 import { licenseSchema } from './license'
 import { notificationSchema } from './notification'
 import { subscriptionSchema } from './subscription'
-import { threadsSchema } from './threads'
 import { mergeSchemas } from './utils'
 import { uuidSchema } from './uuid'
 
@@ -39,7 +38,6 @@ export * from './license'
 export * from './notification'
 export * from './subscription'
 export * from './uuid'
-export * from './threads'
 
 export const schema = mergeSchemas(
   connectionSchema,
@@ -50,6 +48,5 @@ export const schema = mergeSchemas(
   licenseSchema,
   notificationSchema,
   uuidSchema,
-  threadsSchema,
   subscriptionSchema
 )
