--- src/graphql/schema/threads/types.graphql
@@ -1,30 +1,0 @@
-type Thread {
-  createdAt: DateTime!
-  updatedAt: DateTime!
-  title: String!
-  archived: Boolean!
-  trashed: Boolean!
-  object: AbstractUuid!
-  comments(
-    after: String
-    before: String
-    first: Int
-    last: Int
-  ): CommentConnection!
-}
-
-type ThreadsConnection {
-  edges: [ThreadsCursor!]!
-  nodes: [Thread!]!
-  totalCount: Int!
-  pageInfo: PageInfo!
-}
-
-type ThreadsCursor {
-  cursor: String!
-  node: Thread!
-}
-
-type UnsupportedThread {
-  id: Int!
-}
