+++ src/graphql/schema/uuid/thread/index.ts
@@ -25,4 +25,5 @@ import typeDefs from './types.graphql'
 
 export * from './types'
 
+export const threadSchema: Schema = { resolvers, typeDefs: [typeDefs] }
 export const commentSchema: Schema = { resolvers, typeDefs: [typeDefs] }
